scripts/ytt-search/requirements.txt
youtube_transcript_api
python-dotenv
google-api-python-client